requirements.txt
python-dotenv
supabase
httpx
pydantic
jobspy
anthropic
